/getProductsByContainerId  & Añadidod Contenedor.cantidad

+++ src/manager/manager.go
@@ -6,6 +6,8 @@ import (
 	"errors"
 	"strconv"
 	"time"
+
+	"gorm.io/gorm"
 )
 
 type Manager struct {
@@ -56,6 +58,11 @@ func (ma *Manager) CreateProduct(productRequest model.Producto) (model.ProductVi
 		Cantidad:     int16(productRequest.Cantidad),
 	}
 
+	containerUpdatedErr := updateContainer(db, &productRequest, int(productRequest.Cantidad))
+	if containerUpdatedErr != nil {
+		return model.ProductView{}, containerUpdatedErr
+	}
+
 	return product, tx.Error
 }
 
@@ -76,6 +83,13 @@ func (ma *Manager) DeleteProductById(idParam string) error {
 	producto := model.Producto{
 		ID: int32(id),
 	}
+
+	db.Find(&producto)
+	containerUpdatedErr := updateContainer(db, &producto, -int(producto.Cantidad))
+	if containerUpdatedErr != nil {
+		return containerUpdatedErr
+	}
+
 	tx := db.Delete(&producto)
 
 	return tx.Error
@@ -89,6 +103,15 @@ func (ma *Manager) ModifyProduct(productRequestBody model.Producto) (model.Produ
 		return model.ProductView{}, err
 	}
 
+	producto := model.Producto{
+		ID: productRequestBody.ID,
+	}
+	db.Find(&producto)
+
+	if producto.Cantidad != productRequestBody.Cantidad {
+		return model.ProductView{}, errors.New("no está permitido modificar la cantidad")
+	}
+
 	tx := db.Save(&productRequestBody)
 
 	product := model.ProductView{
@@ -98,6 +121,35 @@ func (ma *Manager) ModifyProduct(productRequestBody model.Producto) (model.Produ
 		Cantidad:     int16(productRequestBody.Cantidad)}
 
 	return product, tx.Error
+}
+
+func (ma *Manager) GetProductsByContainerId(idParam string) ([]model.ProductView, error) {
+	db, close, err := db.ObtenerConexionDb()
+	defer close()
+
+	if err != nil {
+		return nil, err
+	}
+
+	id, idParseErr := strconv.Atoi(idParam)
+
+	if idParseErr != nil {
+		return nil, idParseErr
+	}
+
+	productos := []model.Producto{}
+	tx := db.Where("id_contenedor", id).Find(&productos)
+
+	products := []model.ProductView{}
+	for _, producto := range productos {
+		products = append(products, model.ProductView{
+			Id:       producto.ID,
+			Nombre:   producto.Nombre,
+			Cantidad: int16(producto.Cantidad),
+		})
+	}
+
+	return products, tx.Error
 }
 
 func (ma *Manager) GetAllContainers() ([]model.ContainerView, error) {
@@ -114,8 +166,10 @@ func (ma *Manager) GetAllContainers() ([]model.ContainerView, error) {
 	containers := []model.ContainerView{}
 	for _, contenedor := range contenedores {
 		containers = append(containers, model.ContainerView{
-			Id:     contenedor.ID,
-			Nombre: contenedor.Nombre,
+			Id:        contenedor.ID,
+			Nombre:    contenedor.Nombre,
+			Categoria: contenedor.Categoria,
+			Cantidad:  int16(contenedor.Cantidad),
 		})
 	}
 
@@ -133,8 +187,10 @@ func (ma *Manager) CreateContainer(containerRequestBody model.Contenedor) (model
 	tx := db.Create(&containerRequestBody)
 
 	container := model.ContainerView{
-		Id:     containerRequestBody.ID,
-		Nombre: containerRequestBody.Nombre,
+		Id:        containerRequestBody.ID,
+		Nombre:    containerRequestBody.Nombre,
+		Categoria: containerRequestBody.Categoria,
+		Cantidad:  int16(containerRequestBody.Cantidad),
 	}
 
 	return container, tx.Error
@@ -173,8 +229,10 @@ func (ma *Manager) ModifyContainer(containerRequestBody model.Contenedor) (model
 	tx := db.Save(&containerRequestBody)
 
 	container := model.ContainerView{
-		Id:     containerRequestBody.ID,
-		Nombre: containerRequestBody.Nombre,
+		Id:        containerRequestBody.ID,
+		Nombre:    containerRequestBody.Nombre,
+		Categoria: containerRequestBody.Categoria,
+		Cantidad:  int16(containerRequestBody.Cantidad),
 	}
 
 	return container, tx.Error
@@ -228,33 +286,26 @@ func (ma *Manager) AddProductStockById(Idparam string, amountParam string) (mode
 	producto := model.Producto{
 		ID: int32(id),
 	}
-	tx := db.Find(&producto)
 
-	if tx.Error != nil {
-		return model.ProductView{}, tx.Error
+	updatedProduct, productUpdatedErr := updateProduct(db, &producto, amount)
+
+	if productUpdatedErr != nil {
+		return model.ProductView{}, productUpdatedErr
 	}
 
-	productoUpdated := model.Producto{
-		ID:           int32(id),
-		Nombre:       producto.Nombre,
-		IDContenedor: producto.IDContenedor,
-		Cantidad:     producto.Cantidad + int32(amount),
-	}
-	tx = db.Save(&productoUpdated)
+	historyErr := createHistory(producto.ID, int32(amount), "entrada")
 
-	if tx.Error == nil {
-		historyErr := createHistory(productoUpdated.ID, int32(amount), "entrada")
-		if err != nil {
-			return model.ProductView{}, historyErr
-		}
+	if historyErr != nil {
+		return model.ProductView{}, historyErr
 	}
 
-	updatedProduct := model.ProductView{
-		Nombre:   productoUpdated.Nombre,
-		Cantidad: int16(productoUpdated.Cantidad),
+	containerUpdatedErr := updateContainer(db, &producto, amount)
+
+	if containerUpdatedErr != nil {
+		return model.ProductView{}, containerUpdatedErr
 	}
 
-	return updatedProduct, tx.Error
+	return updatedProduct, nil
 }
 
 func (ma *Manager) RemoveProductStockById(Idparam string, amountParam string) (model.ProductView, error) {
@@ -280,30 +331,46 @@ func (ma *Manager) RemoveProductStockById(Idparam string, amountParam string) (m
 	producto := model.Producto{
 		ID: int32(id),
 	}
+	updatedProduct, productUpdatedErr := updateProduct(db, &producto, -amount)
+
+	if productUpdatedErr != nil {
+		return model.ProductView{}, productUpdatedErr
+	}
+
+	historyErr := createHistory(producto.ID, int32(amount), "salida")
+
+	if historyErr != nil {
+		return model.ProductView{}, historyErr
+	}
+
+	containerUpdatedErr := updateContainer(db, &producto, -amount)
+
+	if containerUpdatedErr != nil {
+		return model.ProductView{}, containerUpdatedErr
+	}
+
+	return updatedProduct, nil
+}
+
+func updateProduct(db *gorm.DB, producto *model.Producto, amount int) (model.ProductView, error) {
 	tx := db.Find(&producto)
 
 	if tx.Error != nil {
 		return model.ProductView{}, tx.Error
 	}
 
-	if int32(amount) > producto.Cantidad {
+	productoUpdated := model.Producto{
+		ID:           producto.ID,
+		Nombre:       producto.Nombre,
+		IDContenedor: producto.IDContenedor,
+		Cantidad:     producto.Cantidad + int32(amount),
+	}
+
+	if amount < 0 && int32(amount) < -producto.Cantidad {
 		return model.ProductView{}, errors.New("no hay stock suficiente")
 	}
 
-	productoUpdated := model.Producto{
-		ID:           int32(id),
-		Nombre:       producto.Nombre,
-		IDContenedor: producto.IDContenedor,
-		Cantidad:     producto.Cantidad - int32(amount),
-	}
 	tx = db.Save(&productoUpdated)
-
-	if tx.Error == nil {
-		historyErr := createHistory(productoUpdated.ID, int32(amount), "salida")
-		if err != nil {
-			return model.ProductView{}, historyErr
-		}
-	}
 
 	updatedProduct := model.ProductView{
 		Nombre:   productoUpdated.Nombre,
@@ -311,6 +378,32 @@ func (ma *Manager) RemoveProductStockById(Idparam string, amountParam string) (m
 	}
 
 	return updatedProduct, tx.Error
+}
+
+func updateContainer(db *gorm.DB, producto *model.Producto, amount int) error {
+	contenedor := model.Contenedor{
+		ID: *producto.IDContenedor,
+	}
+
+	tx := db.Find(&contenedor)
+
+	if tx.Error != nil {
+		return tx.Error
+	}
+
+	contenedorUpdated := model.Contenedor{
+		ID:        contenedor.ID,
+		Nombre:    contenedor.Nombre,
+		Categoria: contenedor.Categoria,
+		Cantidad:  contenedor.Cantidad + int32(amount),
+	}
+	tx = db.Save(&contenedorUpdated)
+
+	if tx.Error != nil {
+		return tx.Error
+	}
+
+	return nil
 }
 
 func createHistory(productId int32, amount int32, kind string) error {
